requirements.txt
aldjemy==2.1
asgiref==3.3.4
Django==3.2.4
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.4
greenlet==1.1.0
gunicorn==20.1.0
importlib-metadata==4.5.0
Pillow==8.2.0
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.4.1
